requirements.txt
Django
django-channels
